Schema API

Index

JsonGrinder.ArrayEntry

JsonGrinder.DictEntry

JsonGrinder.LeafEntry

JsonGrinder.Schema

Base.merge

Base.merge!

JsonGrinder.max_string_length

JsonGrinder.max_string_length!

JsonGrinder.max_values

JsonGrinder.max_values!

JsonGrinder.newentry

JsonGrinder.schema

JsonGrinder.update!

API

JsonGrinder.schema — Function

schema([f=identity,] jsons)

Create schema from an array of jsons optionally mapped by function f.

See also: merge, merge!.

JsonGrinder.update! — Function

update!(e, v)

Update the Schema e with value v and return the resulting entry.

JsonGrinder.newentry — Function

newentry(v)

Create and return a new Schema according to the type of v and insert v into it.

Base.merge — Function

merge(schemas...)

Merge multiple schemas into one.

Useful when for example distributing calculation of schema across multiple workers to aggregate all results.

See also: merge!, schema.

Base.merge! — Function

merge!(schema, others...)

Merge multiple schemas others into schema inplace.

See also: merge, schema.

JsonGrinder.Schema — Type

Schema

Supertype for all schema node types.

JsonGrinder.LeafEntry — Type

LeafEntry{T} <: Schema

Keeps statistics about scalar values of type T in leaves.

JsonGrinder.DictEntry — Type

DictEntry <: Schema

Keeps statistics about JSON "objects" containing key-value pairs.

JsonGrinder.ArrayEntry — Type

ArrayEntry <: Schema

Keeps statistics about an "array" entry in JSONs.

JsonGrinder.max_values — Function

JsonGrinder.max_values!(n::Int)

Get the current value of the max_values parameter.

See also: JsonGrinder.max_values!.

JsonGrinder.max_values! — Function

JsonGrinder.max_values!(n::Int; persist=false)

Set the value of the max_values parameter.

Set persist=true to persist this setting between sessions.

See also: JsonGrinder.max_values.

JsonGrinder.max_string_length — Function

JsonGrinder.max_string_length!(n::Int)

Get the current value of the max_string_length parameter.

See also: JsonGrinder.max_string_length!.

JsonGrinder.max_string_length! — Function

JsonGrinder.max_string_length!(n::Int; persist=false)

Set the value of the max_string_length parameter.

Set persist=true to persist this setting between sessions.

See also: JsonGrinder.max_string_length.