API Documentation

JsonGrinder.extractscalar — Method

extractscalar(Type{String}, n = 3, b = 256, m = 2053)

represents strings as ngrams with
	--- `n` (the degree of ngram),
	--- `b` base of string,
	--- `m` modulo on index of the token to reduce dimension

extractscalar(Type{Number}, m = 0, s = 1)
extracts number subtracting `m` and multiplying by `s`

JsonGrinder.generate_html — Method

generate_html(sch::DictEntry; max_vals=100, max_len=1_000)
generate_html(sch::DictEntry, file_name ; max_vals=100, max_len=1_000)

export schema to HTML including CSS style allowing to expand / hide
sub-parts of schema, countmaps, and lengthmaps.

`max_vals` controls maximum number of exported values in countmap
`max_len` controls maximum number of exported lengts of arrays
`file_name` a name of file to save HTML with schema

JsonGrinder.newentry — Method

newentry(v)

	creates new entry describing json according to the type of v

JsonGrinder.prune_json — Method

prune_json(json, schema)

remove keys from json which are not part of the schema

JsonGrinder.schema — Method

function schema(a::Vector{T}) where {T<:Dict}
	function schema(a::Vector{T}) where {T<:AbstractString}
	function schema(map_fun::Function, samples::AbstractArray)

	create schema from an array of parsed or unparsed JSONs

JsonGrinder.suggestextractor — Function

suggestextractor(e::DictEntry, settings = NamedTuple())

	create convertor of json to tree-structure of `DataNode`

	`e` top-level of json hierarchy, typically returned by invoking schema
	`settings.mincount` contains minimum repetition of the key to be included into
	the extractor (if missing it is equal to zero)
	`settings.key_as_field` of the number of keys exceeds this value, it is assumed that
	keys contains a value, which means that they will be treated as strings.
	`settings` can be any container supporting `get` function

JsonGrinder.update! — Method

function update!(a::Entry, v)

	updates the entry when seeing value v

JsonGrinder.updatemaxkeys! — Method

updatemaxkeys!(n::Int)

limits the maximum number of keys in statistics of nodes in JSON. Default value is 10000.

JsonGrinder.updatemaxlen! — Method

updatemaxlen!(n::Int)

limits the maximum size of string values in statistics of nodes in JSON. Default value is 10000.
Longer strings will be trimmed and their length and hash will be appended to retain the uniqueness.
This is due to some strings being very long and causing the schema to be even order of magnitute larger than needed.

JsonGrinder.ArrayEntry — Type

mutable struct ArrayEntry <: JSONEntry
		items
		l::Dict{Int,Int}
		updated::Int
	end

	keeps statistics about an array entry in JSON.
	`items` is typeof `Entry` or nothing and keeps statistics about the elements of the array
	`l` keeps histogram of message length
	`updated` counts how many times the struct was updated.

JsonGrinder.Entry — Type

mutable struct Entry <: JSONEntry
	counts::Dict{Any,Int}
	updated::Int
end

Keeps statistics about scalar values of a one key and also about items inside a key
`counts` counts how many times given value appeared (at most max_keys is held)
`updated` counts how many times the entry was updated

JsonGrinder.ExtractArray — Type

struct ExtractArray{T}
	item::T
end

Convert array of values to a `Mill.BagNode` with items converted 
by `item`. The entire array is assumed to be a single bag.

julia> sc = ExtractArray(ExtractCategorical(Float32,2:4))
julia> sc([2,3,1,4]).data
3×4 Array{Float32,2}:
 1.0  0.0  0.0  0.0
 0.0  1.0  0.0  0.0
 0.0  0.0  0.0  1.0

julia> sc = ExtractArray(ExtractScalar())
julia> sc([2,3,4]).data
 2.0  3.0  4.0

JsonGrinder.ExtractCategorical — Type

ExtractCategorical(s::Entry)
ExtractCategorical(s::UnitRange)
ExtractCategorical(s::Vector)

Converts a single item to a one-hot encoded vector. There is always alocated an extra
element for a unknown value

JsonGrinder.ExtractDict — Type

struct ExtractDict
	dict::Dict{Symbol,Any}
end

extracts all items in `dict` and return them as a ProductNode.

JsonGrinder.ExtractKeyAsField — Type

struct ExtractKeyAsField{S,V} <: AbstractExtractor
	key::S
	item::V
end

extracts all items in `vec` and in `other` and return them as a ProductNode.

JsonGrinder.ExtractOneHot — Type

ExtractOneHot(ks, k, v)

Converts a Vector of `Dict` items to one-hot encoding by using key
`k` to identify a name of item out of `ks` and values
of key `v` as values.

julia> samples = ["{"name": "a", "count" : 1}",
		"{"name": "b", "count" : 2}",];
julia> samples = JSON.parse.(samples);
julia> e = ExtractOneHot(["a","b"], "name", "count");
julia> e(samples).data
3×1 SparseArrays.SparseMatrixCSC{Int64,Int64} with 2 stored entries:
  [1, 1]  =  1
  [2, 1]  =  2

If `v` is equal to `nothing`, then it boils down to one-hot encoding

julia> e = ExtractOneHot(["a","b"], "name", nothing);
julia> e(samples).data
3×1 SparseArrays.SparseMatrixCSC{Int64,Int64} with 2 stored entries:
  [1, 1]  =  1
  [2, 1]  =  1

If there is key in the data which is not known (it was not part of `vs`),
than it is assigned to an special designed key serving as "unknown`

julia> samples = JSON.parse.(["{"name": "c", "count" : 1}"]);
julia> e = ExtractOneHot(["a","b"], "name", nothing);
julia> e(samples).data
3×1 SparseArrays.SparseMatrixCSC{Bool,Int64} with 1 stored entry:
  [3, 1]  =  1

JsonGrinder.ExtractScalar — Type

struct ExtractScalar{T}
	c::T
	s::T
end

Extract a numerical value, centred by subtracting `c` and scaled by multiplying by `s`. 
Strings are converted to numbers. The extractor returnes `ArrayNode{Matrix{T}}` 
with a single row.

JsonGrinder.ExtractString — Type

struct ExtractString{T} <: AbstractExtractor
	n::Int
	b::Int
	m::Int
end

Represent `String` as `n-`grams (`NGramMatrix` from `Mill.jl`) 
with base `b` and modulo `m`.

JsonGrinder.ExtractVector — Type

struct ExtractVector{T}
	item::Int
end

represents an array of a fixed length, typically a feature vector of numbers of type T

julia> sc = ExtractVector(4)
julia> sc([2,3,1,4]).data
3×1 Array{Float32,2}:
 2.0
 3.0
 1.0

JsonGrinder.MultiEntry — Type

mutable struct MultiEntry <: JSONEntry
	childs::Vector{Any}
end

support for JSON which does not adhere to a fixed type.

JsonGrinder.MultipleRepresentation — Type

MultipleRepresentation(extractors::Tuple)

crate a `ProductNode` where each item is the json part
processed by all extractors in the order